# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "SimpleSchedule service API.",
        "title": "SimpleSchedule API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:2811",
    "basePath": "/",
    "paths": {
...
        "/api/tags": {
            "get": {
                "description": "Retrieve a list of all tags",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "tags"
                ],
                "summary": "Get all tags",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handlers.TagsResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/apierror.UserError"
                        }
                    }
                }
            },
            "post": {
                "description": "Create a new tag with title and hex color",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "tags"
                ],
                "summary": "Create tag",
                "parameters": [
                    {
                        "description": "Tag creation payload",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handlers.CreateTagRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/handlers.IDResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/apierror.UserError"
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "$ref": "#/definitions/apierror.UserError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/apierror.UserError"
                        }
                    }
                }
            }
        },
        "/api/tags/{id}": {
            "get": {
                "description": "Retrieve a single tag by its unique identifier",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "tags"
                ],
                "summary": "Get tag by ID",
                "parameters": [
                    {
                        "type": "string",
                        "example": "\"d40c6c2b-e48f-4cb1-80a5-f8c5b6b801a2\"",
                        "description": "Tag ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handlers.TagResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/apierror.UserError"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/apierror.UserError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/apierror.UserError"
                        }
                    }
                }
            },
            "put": {
                "description": "Update title and hex color of an existing tag",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "tags"
                ],
                "summary": "Update tag",
                "parameters": [
                    {
                        "type": "string",
                        "example": "\"d40c6c2b-e48f-4cb1-80a5-f8c5b6b801a2\"",
                        "description": "Tag ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Tag update payload",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handlers.UpdateTagRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handlers.IDResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/apierror.UserError"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/apierror.UserError"
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "$ref": "#/definitions/apierror.UserError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/apierror.UserError"
                        }
                    }
                }
            },
            "delete": {
                "description": "Delete a tag by its ID",
                "tags": [
                    "tags"
                ],
                "summary": "Delete tag",
                "parameters": [
                    {
                        "type": "string",
                        "example": "\"d40c6c2b-e48f-4cb1-80a5-f8c5b6b801a2\"",
                        "description": "Tag ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/apierror.UserError"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/apierror.UserError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/apierror.UserError"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "apierror.UserError": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "message": {
                    "type": "string"
                },
                "timestamp": {
                    "type": "string"
                }
            }
        },
...
        "domain.Tag": {
            "type": "object",
            "properties": {
                "createdAt": {
                    "type": "string"
                },
                "hexColor": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        },
...
        "handlers.CreateTagRequest": {
            "type": "object",
            "required": [
                "hexColor",
                "title"
            ],
            "properties": {
                "hexColor": {
                    "type": "string",
                    "example": "#FF5733"
                },
                "title": {
                    "type": "string",
                    "maxLength": 255,
                    "minLength": 1,
                    "example": "Lecture"
                }
            }
        },
...
        "handlers.IDResponse": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string",
                    "example": "97ec8552-9562-4d79-ab4c-1ca959663629"
                }
            }
        },
...
        "handlers.TagResponse": {
            "type": "object",
            "properties": {
                "tag": {
                    "$ref": "#/definitions/domain.Tag"
                }
            }
        },
        "handlers.TagsResponse": {
            "type": "object",
            "properties": {
                "tags": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/domain.Tag"
                    }
                }
            }
        },
...
        "handlers.UpdateTagRequest": {
            "type": "object",
            "required": [
                "hexColor",
                "title"
            ],
            "properties": {
                "hexColor": {
                    "type": "string",
                    "example": "#33FF57"
                },
                "title": {
                    "type": "string",
                    "maxLength": 255,
                    "minLength": 1,
                    "example": "Practice"
                }
            }
        },
...
    }
}